mail-provisioner/requirements.txt
fastapi
uvicorn[standard]
jinja2
httpx
paramiko
psycopg2-binary
pyyaml
python-dotenv
